* wavetable creation for the saving of waves   * still "issues" with pytables, not sure what's up there
git-svn-id: file:///home/jonas/soma/svn-restore.20080710/software/recorder/trunk@2260 5dccc5d9-beb7-0310-bf71-cc18d937a3f4

+++ src/tspiketable.cc
@@ -61,7 +61,7 @@ TSpikeTable::TSpikeTable(datasource_t src, H5::Group gloc) :
 
   herr_t     status;
 
-  tableName_ = str(boost::format("chan%1d") % src); 
+  tableName_ = str(boost::format("chan%d") % (int)src); 
   
   status = H5TBmake_table( "Table Title", 
 			   tableLoc_.getLocId(),
